pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepagents"
version = "0.1.0"
description = "Product Research Agent using LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Research Agent", email = "ketankr9@gmail.com"}
]
dependencies = [
    "deepagents>=0.4.0",
    "tavily-python>=0.7.0",
    "curl-cffi>=0.7.0",
    "httpx>=0.27.0",
    "markdownify>=0.12.0",
    "beautifulsoup4>=4.12.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "html2text>=2024.2.26",
]

[project.scripts]
product-research = "product_research.cli:main"

[tool.setuptools]
packages = ["product_research"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
